Feat: 포카드인지 확인하는 메소드 추가       * TDD를 도입하려고 했으나 잊고 프로덕션 코드부터 작성했습니다... 	다음 메소드부턴 TDD 방식을 도입하겠습니다!
